Fix TensorArray concatenation, reorg tensor extension tests.
